refactor: adapt to new language
